+++ lib/suspenders/generators/stylesheet_base_generator.rb
@@ -7,7 +7,8 @@ module Suspenders
       File.dirname(__FILE__))
 
     def add_stylesheet_gems
-      gem "bourbon", "5.0.0.beta.6"
+      gem "bourbon", "~> 5.0.0.beta.7"
+      Bundler.with_clean_env { run "bundle install" }
     end
 
     def add_css_config