+++ spec/features/new_project_spec.rb
@@ -47,10 +47,15 @@ RSpec.describe "Suspend a new project with default configuration" do
     end
   end
 
+  it "doesn't generate test directory" do
+    expect(File).not_to exist("#{project_path}/test")
+  end
+
   it "loads secret_key_base from env" do
     secrets_file = IO.read("#{project_path}/config/secrets.yml")
 
-    expect(secrets_file).to match(/secret_key_base: <%= ENV\["SECRET_KEY_BASE"\] %>/)
+    expect(secrets_file).
+      to match(/secret_key_base: <%= ENV\["SECRET_KEY_BASE"\] %>/)
   end
 
   it "adds bin/setup file" do
@@ -124,22 +129,17 @@ RSpec.describe "Suspend a new project with default configuration" do
   it "adds explicit quiet_assets configuration" do
     result = IO.read("#{project_path}/config/application.rb")
 
-    expect(result).to match(
-      /^ +config.quiet_assets = true$/
-    )
+    expect(result).to match(/^ +config.assets.quiet = true$/)
   end
 
-  it "configures static_cache_control in production" do
-    prod_env_file = IO.read("#{project_path}/config/environments/production.rb")
-    expect(prod_env_file).to match(
-      /config.static_cache_control = "public, max-age=.+"/,
+  it "configures public_file_server.headers in production" do
+    expect(production_config).to match(
+      /^ +config.public_file_server.headers = {\n +"Cache-Control" => "public,/,
     )
   end
 
   it "raises on missing translations in development and test" do
-    %w[development test].each do |environment|
-      environment_file =
-        IO.read("#{project_path}/config/environments/#{environment}.rb")
+    [development_config, test_config].each do |environment_file|
       expect(environment_file).to match(
         /^ +config.action_view.raise_on_missing_translations = true$/
       )
@@ -168,9 +168,8 @@ RSpec.describe "Suspend a new project with default configuration" do
   end
 
   it "uses APPLICATION_HOST, not HOST in the production config" do
-    prod_env_file = IO.read("#{project_path}/config/environments/production.rb")
-    expect(prod_env_file).to match(/"APPLICATION_HOST"/)
-    expect(prod_env_file).not_to match(/"HOST"/)
+    expect(production_config).to match(/"APPLICATION_HOST"/)
+    expect(production_config).not_to match(/"HOST"/)
   end
 
   it "configures email interceptor in smtp config" do
@@ -187,7 +186,6 @@ RSpec.describe "Suspend a new project with default configuration" do
 
   it "configs active job queue adapter" do
     application_config = IO.read("#{project_path}/config/application.rb")
-    test_config = IO.read("#{project_path}/config/environments/test.rb")
 
     expect(application_config).to match(
       /^ +config.active_job.queue_adapter = :delayed_job$/
@@ -198,16 +196,12 @@ RSpec.describe "Suspend a new project with default configuration" do
   end
 
   it "configs bullet gem in development" do
-    test_config = IO.read("#{project_path}/config/environments/development.rb")
-
-    expect(test_config).to match /^ +Bullet.enable = true$/
-    expect(test_config).to match /^ +Bullet.bullet_logger = true$/
-    expect(test_config).to match /^ +Bullet.rails_logger = true$/
+    expect(development_config).to match /^ +Bullet.enable = true$/
+    expect(development_config).to match /^ +Bullet.bullet_logger = true$/
+    expect(development_config).to match /^ +Bullet.rails_logger = true$/
   end
 
   it "configs missing assets to raise in test" do
-    test_config = IO.read("#{project_path}/config/environments/test.rb")
-
     expect(test_config).to match(
       /^ +config.assets.raise_runtime_errors = true$/,
     )
@@ -240,9 +234,9 @@ RSpec.describe "Suspend a new project with default configuration" do
     config_files = [
       IO.read("#{project_path}/config/application.rb"),
       IO.read("#{project_path}/config/environment.rb"),
-      IO.read("#{project_path}/config/environments/development.rb"),
-      IO.read("#{project_path}/config/environments/production.rb"),
-      IO.read("#{project_path}/config/environments/test.rb"),
+      development_config,
+      test_config,
+      production_config,
     ]
 
     config_files.each do |file|
@@ -293,15 +287,23 @@ RSpec.describe "Suspend a new project with default configuration" do
     expect(gemfile).to match(/high_voltage/)
   end
 
-  it "adds bourbon" do
-    gemfile = read_project_file("Gemfile")
-
-    expect(gemfile).to match(/bourbon/)
+  it "doesn't use turbolinks" do
+    app_js = read_project_file(%w(app assets javascripts application.js))
+    expect(app_js).not_to match(/turbolinks/)
   end
 
-  it "configures bourbon" do
-    app_css = read_project_file(%w(app assets stylesheets application.sass))
-    expect(app_css).to match(/normalize-rails.*bourbon/m)
+  def development_config
+    @_development_config ||=
+      read_project_file %w(config environments development.rb)
+  end
+
+  def test_config
+    @_test_config ||= read_project_file %w(config environments test.rb)
+  end
+
+  def production_config
+    @_production_config ||=
+      read_project_file %w(config environments production.rb)
   end
 
   def read_project_file(path)